Completion Update
Restructured the PlatformGame class to be more organized and readable, and function with multiple Enemies and Platforms. Also removed unncessary images in the Project files, completed the encoding of the level, and adding a display for the current health of the Player. To do: remove handle edge cases for the boundaries of the window and movement bugs,